Data for the project
This includes the data for the project. Includes the columns mail (has been anonimized, no real emails are provided), the answer text, whether the answer is correct, if the student completed the challenge, and if the answer is the pre or post test. each email appears twice: first for the pre-test, then a post-test answer, this was used to assess learning efficacy of a challenge at Platzi.com, a platform that helps people klearn how to code.